+++ app/src/main/java/woowacourse/payments/ui/component/cardaddition/CardOwnerNameTextField.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.component
+package woowacourse.payments.ui.component.cardaddition
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material3.OutlinedTextField
