+++ app/src/main/java/woowacourse/payments/ui/component/cardaddition/PasswordTextField.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.component
+package woowacourse.payments.ui.component.cardaddition
 
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.OutlinedTextField
