+++ app/src/main/java/woowacourse/payments/ui/component/cardaddition/PaymentCard.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.component
+package woowacourse.payments.ui.component.cardaddition
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
